Add dunst notification manager to i3_wm package
I am not super happy with this but it is definitely part of the window
manager interface. In principle I could setup dunst beforehands but I
want to avoid ordering deps as much as possible. currently the only
hard requirements is shell before bash and xorg before any other GUI stuff.