Add dunst package
Give up on dependency management since it does not actually
matter. The leading concern here is the circular dependency between
the GUI stuff and i3 used for autostart, and xorg with i3. It does not
make sense to install xorg last, so we just give up on the entire idea.