Use eval tricks in init_system and clean_system
Instead of using a case statement on the hosttype variable we use eval
to generate commands that dynamically use the correct package
list. This is kind of dirty but is a usability improvement for the use
case of adding a hosttype and also is very well contained and won't
need further editing. We need to use eval and not indirection
variables as they are not available on macOS system bash.