Stop exporting fish_function_path variable
This doesn't work as expected as the variable is exported to child
processes as a single string instead of an array. This was not necessary
anyway since fish_funtcion_path only makes sense in fish contexts and
config.fish is always executed even for non interactive invocations and
login shells.