[neovim] Add mini.surround plugin
This plugin allows some mappings to modify surrounding pairs, e.g.
quotes.

Unfortunately I am impatient and have a very low timeoutlen set up so
this disables the default keybinding for 's' which is to delete the
character under the cursor and enter insert mode (this behavior is still
accessible using 'cl').

This commit also sets up which-key integration.