[wezterm] Launch with tmux by default
This is done by changing the default program to be a non interactive
login shell that just drops into a tmux session called "scratch"
(creating it if it doesn't exist). If the tmux session is detached it
then drops into an interactive shell...
There is no good standardized way across shell to achieve this behavior
without spawning a sub-shell which is unfortunate. If I could guarantee
fish shell usage then there is the '-C' option which does what I want
[1].

[1] https://github.com/fish-shell/fish-shell/blob/b570c7f6a6a6b9c5c6638d5bc86943342f7d5b5a/doc_src/cmds/fish.rst#L29